change daemonPkg pkg name and other small fixes
daemonPkg is now daemon and the main executable daemon is
service/service.go

Removed incorrect return in server/handler.go when connect
fails

changed some hidden command names per review suggest
