+++ daemon/daemon_test.go
@@ -15,7 +15,7 @@
  *
  */
 
-package daemonPkg
+package daemon
 
 import (
 	"os"
